Change unsigned longs to size_t in SFSpace.

+++ engine/manifold/sfs.cpp
@@ -60,7 +60,7 @@ std::ostream& operator << (std::ostream& out, const SFSFibre& f) {
     return (out << '(' << f.alpha << ',' << f.beta << ')');
 }
 
-SFSFibre SFSpace::fibre(unsigned long which) const {
+SFSFibre SFSpace::fibre(size_t which) const {
     auto pos = fibres_.begin();
     advance(pos, which);
     return *pos;
@@ -166,7 +166,7 @@ void SFSpace::addCrosscap(bool fibreReversing) {
     genus_++;
 }
 
-void SFSpace::addPuncture(bool twisted, unsigned long nPunctures) {
+void SFSpace::addPuncture(bool twisted, size_t nPunctures) {
     if (twisted) {
         puncturesTwisted_ += nPunctures;
 
@@ -196,7 +196,7 @@ void SFSpace::addPuncture(bool twisted, unsigned long nPunctures) {
     }
 }
 
-void SFSpace::addReflector(bool twisted, unsigned long nReflectors) {
+void SFSpace::addReflector(bool twisted, size_t nReflectors) {
     if (twisted) {
         reflectorsTwisted_ += nReflectors;
 
@@ -367,8 +367,8 @@ void SFSpace::reduce(bool mayReflect) {
         // with (p,-q) == (1,-1) (p,p-q) == (p,p-q) if it's worth it.
 
         if (mayReflect) {
-            unsigned long nLarge = 0;
-            unsigned long nSmall = 0;
+            size_t nLarge = 0;
+            size_t nSmall = 0;
             // Don't count (2,1) fibres, they don't get changed anyway.
             for (it = fibres_.begin(); it != fibres_.end() && it->alpha == 2;
                     it++)
@@ -643,8 +643,8 @@ bool SFSpace::operator == (const SFSpace& compare) const {
 std::strong_ordering SFSpace::operator <=> (const SFSpace& rhs) const {
     // Double the genus if it's orientable, so that we can line up tori
     // with Klein bottles, etc.
-    unsigned long adjGenus1 = (baseOrientable() ? genus_ * 2 : genus_);
-    unsigned long adjGenus2 = (rhs.baseOrientable() ?
+    size_t adjGenus1 = (baseOrientable() ? genus_ * 2 : genus_);
+    size_t adjGenus2 = (rhs.baseOrientable() ?
         rhs.genus_ * 2 : rhs.genus_);
 
     // Too many punctures is worse than anything.
@@ -1445,7 +1445,7 @@ AbelianGroup SFSpace::homology() const {
     // abelianise.  The presentation without reflectors is given on
     // p91 of Orlik [1972].  Each reflector gives additional generators
     // y and z, for which y acts as a boundary component and z^2 = fibre.
-    unsigned long nRef = reflectors_ + reflectorsTwisted_;
+    size_t nRef = reflectors_ + reflectorsTwisted_;
     bool twisted = fibreReversing();
 
     if (baseOrientable()) {
@@ -1464,7 +1464,7 @@ AbelianGroup SFSpace::homology() const {
         MatrixInt pres(nFibres_ + nRef + (twisted ? 2 : 1),
             nFibres_ + 1 + 2 * nRef);
 
-        unsigned long which = 0;
+        size_t which = 0;
         for (const auto& f : fibres_) {
             pres.entry(nFibres_ + nRef, which) = 1;
 
@@ -1474,7 +1474,7 @@ AbelianGroup SFSpace::homology() const {
             ++which;
         }
 
-        unsigned long ref;
+        size_t ref;
         for (ref = 0; ref < nRef; ref++) {
             pres.entry(nFibres_ + ref, nFibres_) = -1;
             pres.entry(nFibres_ + ref, nFibres_ + 1 + ref) = 2;
@@ -1506,7 +1506,7 @@ AbelianGroup SFSpace::homology() const {
         MatrixInt pres(nFibres_ + nRef + (twisted ? 2 : 1),
             nFibres_ + genus_ + 1 + 2 * nRef);
 
-        unsigned long which = 0;
+        size_t which = 0;
         for (const auto& f : fibres_) {
             pres.entry(nFibres_ + nRef, which) = 1;
 
@@ -1516,7 +1516,7 @@ AbelianGroup SFSpace::homology() const {
             ++which;
         }
 
-        unsigned long ref;
+        size_t ref;
         for (ref = 0; ref < nRef; ref++) {
             pres.entry(nFibres_ + ref, nFibres_ + genus_) = -1;
             pres.entry(nFibres_ + ref, nFibres_ + genus_ + 1 + ref) = 2;
@@ -1536,7 +1536,7 @@ AbelianGroup SFSpace::homology() const {
     }
 }
 
-void SFSpace::writeBaseExtraCount(std::ostream& out, unsigned long count,
+void SFSpace::writeBaseExtraCount(std::ostream& out, size_t count,
         const char* object, bool tex) {
     out << " + " << count << (tex ? "\\ \\mbox{" : " ") << object;
     if (count != 1)
@@ -1551,8 +1551,8 @@ std::ostream& SFSpace::writeCommonBase(std::ostream& out, bool tex) const {
     // IMPORTANT: We do not allow spaces with > 2 reflector boundary
     // components to be named.  Otherwise this messes up the reflector
     // boundary output.
-    unsigned long totRef = reflectors_ + reflectorsTwisted_;
-    unsigned long totBdries = totRef + punctures_ + puncturesTwisted_;
+    size_t totRef = reflectors_ + reflectorsTwisted_;
+    size_t totBdries = totRef + punctures_ + puncturesTwisted_;
 
     if (baseOrientable()) {
         // Orientable base surface.
